Rework Python modules
