+++ python/cli.schema.json
@@ -78,7 +78,7 @@
           "default": {}
         }
       },
-      "description": "Reference(source_filename: str, filename_md5: str, title: Optional[str] = None, abstract: Optional[str] = None, contents: Optional[str] = None, authors: List[ForwardRef('Author')] = <factory>, chunks: List[ForwardRef('Chunk')] = <factory>, metadata: Dict[str, Any] = <factory>)"
+      "description": "A reference for an academic paper / PDF"
     },
     "Author": {
       "type": "object",
